Implemented mkdir -p and cleaning up superfluous directories.

+++ mfd.c
@@ -176,6 +176,7 @@ static int $mfd_open_sn(
 {
    char fmap[$$PATH_MAX];
    char fdat[$$PATH_MAX];
+   char firstcreated[$$PATH_MAX]; // the first directory created when initialising the map/dat files
    int fd; // map file FD
    int fd_dat; // dat file FD
    int ret;
@@ -201,9 +202,16 @@ static int $mfd_open_sn(
    // Get the paths of the map & dat files
    $$ADDNPSFIX_CONT(fmap, fdat, vpath, fsdata->sn_lat_dir, fsdata->sn_lat_dir_len, $$EXT_MAP, $$EXT_DAT, $$EXT_LEN)
 
-   //////////////////////////////////
-   // TODO implement mkdir -p here // TODO which creates directories mirroring them in the main space
-   ////////////////////////////////// (create them when THEY are modified)
+   // Create path
+   ret = $mkpath(fmap, firstcreated, S_IRWXU);
+   if(ret < 0){ // error
+      $dlogdbg("mfd_open_sn: mkpath failed with %d = %s\n", -ret, strerror(-ret));
+      return ret;
+   }
+   if(ret == 0){ // no directory needed to be created
+      firstcreated[0] = '\0';
+   }
+   $dlogdbg("mfd_open_sn: mkpath created directories under '%s' (or none)\n", firstcreated);
 
    // Open or create the map file
    fd = open(fmap, O_RDWR | O_CREAT | O_EXCL, S_IRWXU); // TODO add O_NOATIME
@@ -329,7 +337,13 @@ static int $mfd_open_sn(
 
       if(waserror != 0){ // Cleanup
          close(fd);
-         unlink(fmap);
+         if(firstcreated[0] == '\0'){
+            $dlogdbg("mfd_open_sn: cleanup: removing file %s\n", fmap);
+            unlink(fmap);
+         } else {
+            $dlogdbg("mfd_open_sn: cleanup: removing directory %s\n", firstcreated);
+            $recursive_remove(firstcreated);
+         }
          return -waserror;
       }
